[dawn][native] Spanify CommandIterator::NextData
And also make it return a const span to ensure users don't modify the
command data.

This requires modifying every backend's command iteration to retrieve
Spans instead ot T* from NextData and either unpack it again with
.data() and .size(), or pass the Span further.

Change a couple of sizes in Commands.h to be size_t so that a Span<> is
returned from NextData when that size's type is used.

Bug: 528305452, 532944732, 532946455
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/323079
Commit-Queue: Corentin Wallez <cwallez@chromium.org>

+++ src/dawn/native/Commands.h
@@ -411,7 +411,8 @@ struct SetBindGroupCmd {
 
     BindGroupIndex index = BindGroupIndex(0u);
     Ref<BindGroupBase> group;
-    uint32_t dynamicOffsetCount = 0;
+    // TODO(https://crbug.com/532944732) Use BindingIndex.
+    size_t dynamicOffsetCount = 0;
 };
 
 struct SetImmediatesCmd {
@@ -419,7 +420,7 @@ struct SetImmediatesCmd {
     ~SetImmediatesCmd();
 
     uint32_t offset = 0;
-    uint32_t size = 0;
+    size_t size = 0;
 };
 
 struct SetIndexBufferCmd {
@@ -455,7 +456,7 @@ struct WriteBufferCmd {
 
     Ref<BufferBase> buffer;
     uint64_t offset = 0;
-    uint64_t size = 0;
+    size_t size = 0;
 };
 
 struct WriteTimestampCmd {
